Extract `replica_set/read_ops.rs`
